Implemented a strategy to decide if sampling new point or re-sampling
The strategy is based on comparing the loss of the interval with the loss of the same interval but taking different 'y' values. These values are the extremes of the interval of confidence of the true value of the mean. This method does not seem good after several tests on peak and sinus functions. It focuses too strongly on regions with large gradients.