Added 'strategy' input to AverageLearner1D and coded strategy 2
The input 'strategy' determines which strategy is used to decide whether sampling a new point or re-sampling. These strategies are explained in the slides for the meeting on 17 April. This is only a testing tool, since the final version of the code will only have one strategy implemented and therefore no 'strategy' input.